backend/pyproject.toml
[project]
name = "hank-backend"
version = "0.1.0"
description = "Hank — a voice AI tutor for house maintenance"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[openai,silero]~=1.4",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "livekit-api>=0.8.0",
    "openai>=1.50.0",
]

[tool.uv]
package = false
